Add files via upload
English translation:

The question is as folllows :
Suppose two people are standing at a distance from each other. If we observe these people from above, we can imagine the surface on which they are standing as a set of rows and columns. In fact, this surface can be considered as a coordinate plane with rows parallel to the horizontal axis and columns parallel to the vertical axis. Write a program that takes the coordinates of these two people, including the row and column, from the user and outputs whether the line of distance between these two people is horizontal or parallel.